Do not use networkclient in compute package
